# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: zekurix-server
  description: Backend REST API for the Zekurix platform
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  version: 0.1.0
paths:
  /health:
    get:
      tags:
        - health
      summary: Health check endpoint.
      description: Returns the current application status and version.
      operationId: get_health
      responses:
        "200":
          description: Current application status and version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /users:
    post:
      tags:
        - users
      summary: Create a new user.
      description: Creates a new user with the provided username and returns the
        created user information.
      operationId: create_user
      requestBody:
        description: The user parameters required to create the user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "409":
          description: A user with this username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: The username is invalid. It must be between 3 and 64 characters and
            may only contain ASCII alphanumeric characters, '_' and '-'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: An unexpected internal server error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      tags:
        - users
      summary: Retrieve a user by ID.
      description: Returns the user information for the specified user ID.
      operationId: get_user
      parameters:
        - name: id
          in: path
          description: The unique identifier of the user
          required: true
          schema:
            $ref: "#/components/schemas/UserId"
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: An unexpected internal server error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    CreateUserRequest:
      type: object
      description: Request parameters for creating a new user.
      required:
        - username
      properties:
        username:
          type: string
          description: The username for the new user.
          example: Alice
    ErrorResponse:
      type: object
      required:
        - code
      properties:
        code:
          type: string
    HealthResponse:
      type: object
      description: Response returned by the health check endpoint.
      required:
        - status
        - version
      properties:
        status:
          type: string
          description: Current health status of the application.
          example: ok
        version:
          type: string
          description: Current application version.
          example: 1.0.0
    UserId:
      type: string
      format: uuid
    UserResponse:
      type: object
      description: Response containing user information.
      required:
        - id
        - username
      properties:
        id:
          $ref: "#/components/schemas/UserId"
          description: The unique identifier of the user.
        username:
          $ref: "#/components/schemas/Username"
          description: The username of the user.
    Username:
      type: string
      example: alice-123
      pattern: ^[A-Za-z0-9_-]{3,64}$